Read scanimage tiffs into an R object

Read scanimage tiffs into an R object

read.scanimage(source, slices = Inf, frames = Inf, channels = Inf,
  info = T, as.is = TRUE, ...)

Arguments

source
Either name of the file to read from or a raw vector representing the TIFF file content.

slices
Which slices to return (default of Inf reads all available slices). A slice is a single channel 2D TIFF image.

frames
Which frames to read (default of Inf reads all available frames). A frame is one or more 2D TIFF images acquired simultaneously.

channels
which channels to read (default value of Inf implies all channels).

info
if set to TRUE then the resulting image(s) will also contain information from TIFF tags as attributes

as.is
Return original pixel values without rescaling where possible. This is usually what you want for 8 and 16 bit scanimage TIFFs (default TRUE).

...
additional arguments for readTIFF.

Value

A list of 2D matrices (height x width) except for the special case when the first slice has been requested.

Examples

t=read.scanimage(system.file('extdata/Blank-IPA_1s_16r_032.tif', package='scanimage'))